# docs/api/storages-swagger-api.yaml
openapi: 3.0.0

info:
    version: v1
    title: API для работы с пунктами хранения в сервисе по бронированию камер хранения

tags:
    -   name: storages
        description: Пункты хранения

paths:
    /api/v1/storages:
        get:
            tags:
                - storages
            summary: Найти пункты хранения
            parameters:
                -   name: name
                    in: query
                    required: false
                    description: Название пункта хранения
                    schema:
                        type: string
                        example: Хранилище "У Рязановой"
                -   name: address
                    in: query
                    required: false
                    description: Адрес пункта хранения
                    schema:
                        type: string
                        example: Москва
                -   name: pageNumber
                    in: query
                    required: false
                    description: Номер страницы
                    schema:
                        type: integer
                        format: int32
                        example: 0
                -   name: pageSize
                    in: query
                    required: false
                    description: Размер страницы
                    schema:
                        type: integer
                        format: int32
                        example: 50
            responses:
                '200':
                    description: Данные получены
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    $ref: '#/components/schemas/StorageGet'
                '400':
                    description: Некорректные параметры поиска
                '500':
                    description: Ошибки сервера
        post:
            tags:
                - storages
            summary: Создать пункт хранения
            requestBody:
                description: Форма для создания пункта хранения
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/StorageCreate'
            responses:
                '204':
                    description: Пункт хранения создан
                '400':
                    description: Некорректные данные формы создания пункта хранения
                '405':
                    description: Сервер запущен только на чтение
                '409':
                    description: Пользователь с таким логином или почтой уже существует
                '500':
                    description: Ошибки сервера
    /api/v1/storages/requests:
        get:
            tags:
                - storages
            summary: Найти заявки на создание пунктов хранения
            security:
                -   JWTBearer: [ ]
            parameters:
                -   name: name
                    in: query
                    required: false
                    description: Название пункта хранения
                    schema:
                        type: string
                        example: Хранилище "У Рязановой"
                -   name: address
                    in: query
                    required: false
                    description: Адрес пункта хранения
                    schema:
                        type: string
                        example: Москва
                -   name: pageNumber
                    in: query
                    required: false
                    description: Номер страницы
                    schema:
                        type: integer
                        format: int32
                        example: 0
                -   name: pageSize
                    in: query
                    required: false
                    description: Размер страницы
                    schema:
                        type: integer
                        format: int32
                        example: 50
            responses:
                '200':
                    description: Данные получены
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    $ref: '#/components/schemas/StorageGet'
                '400':
                    description: Некорректные параметры поиска
                '401':
                    description: Пользователь не авторизован
                '403':
                    description: Пользователь не имеет прав на эту операцию
                '500':
                    description: Ошибки сервера
    /api/v1/storages/{storageId}:
        get:
            tags:
                - storages
            summary: Просмотреть данные пункта хранения
            security:
                -   JWTBearer: [ ]
            parameters:
                -   name: storageId
                    in: path
                    required: true
                    description: Идентификатор пункта хранения
                    schema:
                        type: string
                        format: uuid
            responses:
                '200':
                    description: Данные получены
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/StorageGet'
                '400':
                    description: Некорректные параметры запроса
                '404':
                    description: Пункт хранения не найден
                '500':
                    description: Ошибки сервера
        patch:
            tags:
                - storages
            summary: Обновить данные пункта хранения
            security:
                -   JWTBearer: [ ]
            parameters:
                -   name: storageId
                    in: path
                    required: true
                    description: Идентификатор пункта хранения
                    schema:
                        type: string
                        format: uuid
            requestBody:
                description: Форма для обновления пункта хранения
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/StorageUpdate'
            responses:
                '200':
                    description: Данные обновлены
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/StorageGet'
                '400':
                    description: Некорректные данные формы обновления
                '401':
                    description: Пользователь не авторизован
                '403':
                    description: Пользователь не имеет прав на эту операцию
                '404':
                    description: Пользователь не найден
                '405':
                    description: Сервер запущен только на чтение
                '500':
                    description: Ошибки сервера
    /api/v1/storages/{storageId}/approve:
        patch:
            tags:
                - storages
            summary: Подтвердить данные пункта хранения
            security:
                -   JWTBearer: [ ]
            parameters:
                -   name: storageId
                    in: path
                    required: true
                    description: Идентификатор пункта хранения
                    schema:
                        type: string
                        format: uuid
            responses:
                '200':
                    description: Пункт хранения подтверждён
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/StorageGet'
                '400':
                    description: Некорректные параметры запроса
                '401':
                    description: Пользователь не авторизован
                '403':
                    description: Пользователь не имеет прав на эту операцию
                '404':
                    description: Пункт хранения не найден
                '405':
                    description: Сервер запущен только на чтение
                '500':
                    description: Ошибки сервера
    /api/v1/storages/{storageId}/reject:
        delete:
            tags:
                - storages
            summary: Отклонить заявку на создание пункта хранения
            security:
                -   JWTBearer: [ ]
            parameters:
                -   name: storageId
                    in: path
                    required: true
                    description: Идентификатор пункта хранения
                    schema:
                        type: string
                        format: uuid
            responses:
                '204':
                    description: Заявка отклонена
                '400':
                    description: Некорректные параметры запроса
                '401':
                    description: Пользователь не авторизован
                '403':
                    description: Пользователь не имеет прав на эту операцию
                '404':
                    description: Пункт хранения не найден
                '405':
                    description: Сервер запущен только на чтение
                '500':
                    description: Ошибки сервера
    /api/v1/storages/my:
        get:
            tags:
                - storages
            summary: Просмотреть данные пункта хранения
            security:
                -   JWTBearer: [ ]
            responses:
                '200':
                    description: Данные получены
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/StorageGet'
                '401':
                    description: Пользователь не авторизован
                '403':
                    description: Пользователь не имеет прав на эту операцию
                '500':
                    description: Ошибки сервера
        patch:
            tags:
                - storages
            summary: Обновить данные пункта хранения
            security:
                -   JWTBearer: [ ]
            requestBody:
                description: Форма для обновления пункта хранения
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/StorageUpdate'
            responses:
                '200':
                    description: Данные обновлены
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/StorageGet'
                '400':
                    description: Некорректные данные формы обновления
                '401':
                    description: Пользователь не авторизован
                '403':
                    description: Пользователь не имеет прав на эту операцию
                '404':
                    description: Пользователь не найден
                '405':
                    description: Сервер запущен только на чтение
                '500':
                    description: Ошибки сервера

components:
    schemas:
        StorageCreate:
            $ref: '../dtos/storages/storage-create.schema.yaml'
        StorageGet:
            $ref: '../dtos/storages/storage-get.schema.yaml'
        StorageStatus:
            $ref: '../dtos/storages/storage-status.schema.yaml'
        StorageUpdate:
            $ref: '../dtos/storages/storage-update.schema.yaml'
    securitySchemes:
        JWTBearer:
            $ref: '../security/jwt.schema.yaml'